+++ js/filter.js
@@ -1,0 +1,1 @@
+// 5.8. Объекты, расположенные неподалёку, можно фильтровать.
