+++ js/form.js
@@ -1,0 +1,7 @@
+// 3. Ограничения, накладываемые на поля ввода
+// ?Возможно соединить с модулем validation.js, т.к. в нём происходит проверка на валидность заполненных данных
+
+// function submitForm () {
+//   Если отправка данных прошла успешно, показывается соответствующее сообщение.
+//   Если при отправке данных произошла ошибка запроса, показывается соответствующее сообщение
+// }
