+++ js/map.js
@@ -1,0 +1,1 @@
+// 4. Выбор адреса на карте
