+++ js/preview.js
@@ -1,0 +1,2 @@
+// 6.1. В форме подачи объявления показывается превью аватарки пользователя
+// и фотографии помещения при изменении значений соответствующих полей.
