+++ js/similar.js
@@ -1,0 +1,1 @@
+// 5. Сравнение с похожими объявлениями
