+++ js/state.js
@@ -1,0 +1,4 @@
+// 1. Состояния страницы
+
+// function activeState () {}
+// function inactiveState () {}
