+++ js/submit.js
@@ -1,0 +1,9 @@
+// 2.6. Если отправка данных прошла успешно, показывается соответствующее сообщение.
+// Сообщение должно исчезать по нажатию на клавишу Esc и по клику на произвольную область экрана.
+
+// function submitSuccessful () {}
+
+// 2.7. Если при отправке данных произошла ошибка запроса, показывается соответствующее сообщение.
+// Сообщение должно исчезать после нажатия на кнопку .error__button, по нажатию на клавишу Esc и по клику на произвольную область экрана
+
+// function submitFailed () {}
