+++ js/validation.js
@@ -1,0 +1,12 @@
+// 2.4. Страница реагирует на неправильно введённые значения в форму.
+
+// При попытке отправить форму с неправильными данными, отправки не происходит,
+// а неверно заполненные поля подсвечиваются красной рамкой.
+
+// function makeValidation () {}
+
+// 2.5. При успешной отправке формы или её очистке...
+// ?либо отдельная функция
+// function resetForm () {
+// }
+// ?либо часть makeValidation
